si el programa no recibe inputs incorrectos, la memoria se libera

+++ inc/input.h
@@ -1,6 +1,6 @@
 #ifndef INPUT_H
 #define INPUT_H
 
-int load_input_as_coord(float***, float*);
+int load_input_as_coord(float***, int*);
 
 #endif
